Exits with exit code of 1 on any error
This commit replaces all existing exception handling to allow errors to flow
up to the global handler.  This ensures that any error will get caught by the
global error handler and cause bundler to exit with a non 0 exit code.  This
causes a problem with the VS extension because it is not correctly picking
up the writes to stderr.  The bundler extension still does everything else
correctly.